refactor: FoodName에서 menu를 반환하는 역할을 Menu로 위임
