+++ src/main/java/christmas/domain/event/EventCondition.java
@@ -1,5 +1,6 @@
 package christmas.domain.event;
 
+import christmas.domain.food.Menu;
 import christmas.domain.order.Orders;
 
 public class EventCondition {
@@ -7,13 +8,13 @@ public class EventCondition {
 
     public static boolean isOrderPricesAboveThreshold(Orders orders) {
         return orders.getOrders().stream()
-                .mapToInt(order -> order.menu().getPrice() * order.getQuantity())
+                .mapToInt(order -> Menu.from(order.getName()).getPrice() * order.getQuantity())
                 .sum() >= EVENT_COMMON_THRESOLD;
     }
 
     public static boolean isOrderPricesAboveThreshold(Orders orders, long thresold) {
         return orders.getOrders().stream()
-                .mapToInt(order -> order.menu().getPrice() * order.getQuantity())
+                .mapToInt(order -> Menu.from(order.getName()).getPrice() * order.getQuantity())
                 .sum() >= thresold;
     }
 }
