+++ src/main/java/christmas/domain/event/policy/gift/FoodGiftPolicy.java
@@ -1,8 +1,8 @@
 package christmas.domain.event.policy.gift;
 
-import christmas.domain.food.FoodItem;
 import christmas.domain.event.EventCondition;
 import christmas.domain.order.Orders;
+import christmas.dto.FoodItem;
 import java.util.Optional;
 
 public class FoodGiftPolicy implements GiftPolicy {
