+++ src/main/java/christmas/domain/order/Orders.java
@@ -1,6 +1,6 @@
 package christmas.domain.order;
 
-import christmas.domain.food.FoodItem;
+import christmas.dto.FoodItem;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
