+++ src/main/java/christmas/domain/price/PriceCalculator.java
@@ -1,19 +1,20 @@
 package christmas.domain.price;
 
-import christmas.domain.food.FoodItem;
+import christmas.domain.food.Menu;
+import christmas.dto.FoodItem;
 import java.util.List;
 
 //TODO 역할 분리
 public class PriceCalculator {
     public int calculateItemsPrice(List<FoodItem> items) {
         return items.stream()
-                .mapToInt(item -> item.menu().getPrice() * item.getQuantity())
+                .mapToInt(item -> Menu.from(item.getName()).getPrice() * item.getQuantity())
                 .sum();
     }
 
     public int calculateTotalGifts(List<FoodItem> items) {
         return items.stream()
-                .mapToInt(item -> item.menu().getPrice() * item.getQuantity())
+                .mapToInt(item -> Menu.from(item.getName()).getPrice() * item.getQuantity())
                 .sum();
     }
