+++ src/main/java/christmas/service/EventService.java
@@ -3,11 +3,11 @@ package christmas.service;
 import christmas.domain.event.policy.PolicyManager;
 import christmas.domain.event.policy.discount.DiscountPolicy;
 import christmas.domain.event.policy.gift.GiftPolicy;
-import christmas.domain.food.FoodItem;
 import christmas.domain.order.Orders;
 import christmas.domain.price.PriceCalculator;
 import christmas.domain.price.money.KoreaMoney;
 import christmas.domain.price.money.Money;
+import christmas.dto.FoodItem;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
