+++ src/test/java/christmas/domain/FoodItemTest.java
@@ -1,11 +1,10 @@
 package christmas.domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.internal.bytebuddy.matcher.ElementMatchers.is;
-import static org.junit.jupiter.api.Assertions.*;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-import christmas.domain.food.FoodItem;
+import christmas.dto.FoodItem;
 import org.junit.jupiter.api.Test;
 
 class FoodItemTest {
